+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 [Unreleased]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-3.14.6...HEAD
 
+### Added
+
+- #1229 - Added config option to remove trailing slash from extensionless URLs in sitemap.
+
 ## [3.14.6] - 2018-01-31
 
 ### Fixed
@@ -34,7 +38,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 - #989 - Dynamic Loading for optional Touch UI ClientLibraries.
 - #1218 - New Report Builder Feature.
-- #1228 - Added config option to have extensionless URLs in sitemap
+- #1228 - Added config option to have extensionless URLs in sitemap.
 
 ### Changed